+++ src/lib/components/forms/select-box/terra-select-box.component.spec.ts
@@ -3,7 +3,6 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
 import { LocalizationModule } from 'angular-l10n';
 import { l10nConfig } from '../../../../app/translation/l10n.config';
-import { TerraLabelTooltipDirective } from '../../../helpers/terra-label-tooltip.directive';
 import { TerraSelectBoxValueInterface } from './data/terra-select-box.interface';
 import { TooltipDirective } from '../../tooltip/tooltip.directive';
 import { By } from '@angular/platform-browser';
@@ -40,7 +39,7 @@ describe('TerraSelectBoxComponent:', () => {
 
     beforeEach(() => {
         TestBed.configureTestingModule({
-            declarations: [TooltipDirective, TerraSelectBoxComponent, TerraLabelTooltipDirective],
+            declarations: [TooltipDirective, TerraSelectBoxComponent],
             imports: [FormsModule, LocalizationModule.forRoot(l10nConfig)],
             providers: [
                 {